requirements.txt
Django==2.1.3
djangorestframework==3.9.0
djangorestframework-simplejwt==3.3
PyJWT==1.7.1
PyMySQL==0.9.3
pytz==2018.9
